dom/xhr/tests/test_XHR_header.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for XMLHttpRequest.GetResponseHeader(foo) byte-inflates the output</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta charset="utf-8">
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.7">
"use strict";
SimpleTest.waitForExplicitFinish();

let xhr = new XMLHttpRequest();
xhr.open('GET', 'file_XHR_header.sjs', true);
xhr.onreadystatechange = function() {
  if (xhr.readyState == 4) {
    ok(xhr.getResponseHeader('X-Custom-Header-Bytes') == "\xE2\x80\xA6", 'getResponseHeader byte-inflates the output');
    SimpleTest.finish();
  }
}
xhr.send(null);
</script>
</pre>
</body>
</html>
